function_app/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pyjwt
requests
jmespath
aiohttp

pandas
streamlit

azure-functions
azure-core
azure-identity
azure-mgmt-cognitiveservices
azure-storage-blob
azure-mgmt-web
